Moved get_voice_sfx_choices to Sounds.py where it belongs with its friends
