+++ SettingsList.py
@@ -1673,25 +1673,6 @@ logic_tricks = {
 }
 
 
-def get_voice_sfx_choices(age, include_random=True):
-    # Dynamically populate the SettingsList entry for the voice effects
-    # Voice packs should be a folder of .bin files in the Voices directory
-    # The first 5 characters should be either Child or Adult
-    names = ['Default', 'Silent']
-    sfxpath = data_path('Voices')
-    sfxkey = 'Child' if age == 0 else 'Adult'
-    if os.path.exists(sfxpath):
-        for file in os.listdir(sfxpath):
-            if os.path.isdir(os.path.join(sfxpath, file)):
-                if file[:5] == sfxkey:
-                    names.append(file[5:])
-
-    # Add a random if multiple options are available
-    if len(names) > 2 and include_random:
-        names.append('Random')
-    return names
-
-
 # a list of the possible settings
 setting_infos = [
     # Web Only Settings
@@ -6015,7 +5996,7 @@ setting_infos = [
         gui_text       = 'Adult Voice',
         shared         = False,
         cosmetic       = True,
-        choices        = get_voice_sfx_choices(1),
+        choices        = sfx.get_voice_sfx_choices(1),
         default        = 'Default',
         gui_tooltip    = '''\
             Change Link's adult voice.
@@ -6029,7 +6010,7 @@ setting_infos = [
         gui_text       = 'Child Voice',
         shared         = False,
         cosmetic       = True,
-        choices        = get_voice_sfx_choices(0),
+        choices        = sfx.get_voice_sfx_choices(0),
         default        = 'Default',
         gui_tooltip    = '''\
             Change Link's child voice.
