+++ Sounds.py
@@ -26,6 +26,8 @@
 
 from enum import Enum
 from collections import namedtuple
+import os
+from Utils import data_path
 
 
 class Tags(Enum):
@@ -210,3 +212,22 @@ def get_setting_choices(sound_hook):
         **choices,
         }
     return result
+
+
+def get_voice_sfx_choices(age, include_random=True):
+    # Dynamically populate the SettingsList entry for the voice effects
+    # Voice packs should be a folder of .bin files in the Voices directory
+    # The first 5 characters should be either Child or Adult
+    names = ['Default', 'Silent']
+    sfxpath = data_path('Voices')
+    sfxkey = 'Child' if age == 0 else 'Adult'
+    if os.path.exists(sfxpath):
+        for file in os.listdir(sfxpath):
+            if os.path.isdir(os.path.join(sfxpath, file)):
+                if file[:5] == sfxkey:
+                    names.append(file[5:])
+
+    # Add a random if multiple options are available
+    if len(names) > 2 and include_random:
+        names.append('Random')
+    return names
